% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calcMZs.R
\name{calcIons}
\alias{calcIons}
\title{calcIons}
\usage{
calcIons(MF, charges = c(1), carrier = "H", monoisotopic = T,
  mf_column = "formula", adduct = NULL)
}
\arguments{
\item{MF}{character vector with molecular formulas (MFs)}

\item{charges}{vector of positive or negative integers indicating the charge states to calculate}

\item{carrier}{charge carrier - Molecular formula will be added to or removed from MFs for each charge}

\item{monoisotopic}{if true, will calculate the monoisotopic mass m/z values, 
will calculate m/z of the most abundant isotope peak otherwise (preferrable e.g. for large organic molecules)}

\item{mf_column}{name of column in result data.frame that will contain the molecular formulas}

\item{adduct}{an adduct, e.g. "Na" or "Cl". If the charge is positive, each adduct replaces one charge carrier.}
}
\value{
a data.frame with columns \code{mz}, \code{charge}, \code{ion} and the \code{mf_column} as specified
}
\description{
Calculate charge-to-mass ratios for charge states of ions
}
