% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MF_prediction.R
\name{calcMF}
\alias{calcMF}
\title{calcMF}
\usage{
calcMF(mz = 200.000659, z = 1, ppm = 5, top = NULL,
  elements = Rdisop::initializeCHNOPS(), maxCounts = TRUE,
  SeniorRule = TRUE, HCratio = TRUE, moreRatios = TRUE,
  elementHeuristic = TRUE, Filters = list(DBErange = c(-5, 40),
  minElements = "C0", maxElements = "C99999", parity = "e", maxCounts =
  TRUE, SENIOR3 = 0, HCratio = TRUE, moreRatios = TRUE, elementHeuristic =
  TRUE), summarize = FALSE, BPPARAM = NULL)
}
\arguments{
\item{mz}{monoisotopic mass or monoisotopic m/z value}

\item{z}{charge (set 0 for uncharged)}

\item{ppm}{error tolerance in ppm}

\item{top}{if not NULL, maximum number of \code{top} formulas to be returned, as ranked by ppm error}

\item{elements}{either NULL or an element list for \pkg{Rdisop}, as generated by \link[Rdisop]{initializeCHNOPS} and related functions.}

\item{maxCounts}{check each predicted molecula formula for maximum number of elements expected in a natural product of its size ("Golden Rule #1")}

\item{SeniorRule}{for each predicted molecula formula, calculate the sum of valences minus twice (the number of atoms minus 1) (to check for Senior's third theorem, "Golden Rule #2")[1][2][3]}

\item{HCratio}{for each predicted molecula formula, calculate Hydrogen/ Carbon count ratio ("Golden Rule #4")}

\item{moreRatios}{for each predicted molecula formula, calculate Nitrogen/Carbon, Oxygen/Carbon, Phosphorus/Carbon and Sulfur/Carbon count ratios ("Golden Rule #5")}

\item{elementHeuristic}{apply a heuristic HNOPS probability check as proposed in "Golden Rule #6"}

\item{Filters}{which filters to apply, see Details}

\item{summarize}{if \code{TRUE}, will return a single character string with molecular formulas and charge instead of a data.table}

\item{BPPARAM}{parameters for parallel processing with \pkg{BiocParallel}. Can be \link[BiocParallel]{bpparam}(), or an object of these classes \link[BiocParallel]{SnowParam-class}, \link[BiocParallel]{MulticoreParam-class}. If NULL, \link[BiocParallel]{SerialParam}() is called (no multicore processing).}
}
\value{
a data.frame (or list of data.frames if multiple mz values are supplied) with molecular formulas generated with \link[Rdisop]{decomposeMass}() with additional information 
and optionally filtered as described. NOTE: If \code{summarize = TRUE}, results for each query mz value are summarized in a single character string with all molecular formulas matching filter criteria
}
\description{
Calculate molecular formulas from a monoisotopic mass. This function uses \pkg{Rdisop} to generate molecular formulas, and can apply many of the filters described as the "seven golden rules"[1]
}
\details{
\subsection{Filters}{
These filtering steps can be selected. Filter items that are NULL (or not in the list) will not be applied.

\describe{
   \item{DBErange}{numeric vector of length 2, defining lower and upper limit for double bond equivalents in the molecule}
   \item{minElements}{character(1) with a molecula formula defining the minimum number of atom counts for a set of elements}
   \item{maxElements}{character(1) with a molecula formula defining the maximum number of atom counts for a set of elements}
   \item{parity}{filter for parity. must be \code{"e"} (for even) or \code{"o"} (for odd). What to expect depends on ionization technique used.}
   \item{maxCounts}{if \code{TRUE}, uses a filter limiting atom counts for elements as described by Kind & Fiehn [1]}
   \item{SENIOR3}{If \code{TRUE}, the sum of valences has to be greater than or equal to twice the number of atoms minus 1, as Senior's third theorem suggests [1][2][3]}
   \item{HCratio}{If \code{TRUE}, molecular formulas must have a Hydrogen/ Carbon count ratio of >0.2 and <3.1 ("Golden Rule #4")[1]. Will be ignored if no Hydrogen or Carbon present in a molecule.}
   \item{moreRatios}{If \code{TRUE}, molecular formulas must have these atom count ratios: Nitrogen/Carbon <1.3, Oxygen/Carbon <1.2, Phosphorus/Carbon <0.3 and Sulfur/Carbon <0.8 ("Golden Rule #5")[1]. Will be ignored if no Carbon present in a molecule.}
   \item{elementHeuristic}{If \code{TRUE}, additional element ratio heuristic is applied ("Golden Rule #6") [1]}
}}
}
\examples{

calcMF(mz = 200.000659, z = 1, ppm = 5)

}
\references{
\enumerate{
\item Kind T, Fiehn O (2007) Seven Golden Rules for heuristic filtering of molecular formulas obtained by accurate mass spectrometry. BMC Bioinformatics 8:105. doi: 10.1186/1471-2105-8-105
\item Senior JK (1951) Partitions and Their Representative Graphs. Am J Math 73:663. doi: 10.2307/2372318
\item Morikawa T, Newbold BT (2003) Analogous odd-even parities in mathematics and chemistry. Chemistry 12:445–450
}
}
