% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_chemical_formula_funcitons.R
\name{calculate_mass}
\alias{calculate_mass}
\title{Calculate the Mass of a Molecular Formula}
\usage{
calculate_mass(formula, which = c("exact_mass", "average_mass"))
}
\arguments{
\item{formula}{Character string containing a molecular formula.}

\item{which}{Character string specifying whether to calculate `"exact_mass"`
or `"average_mass"`.}
}
\value{
A named numeric value giving the calculated mass.
}
\description{
Calculate the exact or average mass of a molecular formula.
}
\details{
Calculate the Mass of a Molecular Formula

Calculate the exact or average mass of a molecular formula from its element
composition.
}
\examples{
calculate_mass("H2O")                      # Using default exact_mass
calculate_mass("H2O", which = "average_mass")
calculate_mass("C6H12O6", which = "exact_mass")

}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
