% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_check_functions.R
\name{check_adduct_formula}
\alias{check_adduct_formula}
\title{Validate an Adduct Formula}
\usage{
check_adduct_formula(adduct)
}
\arguments{
\item{adduct}{A character string to validate.}
}
\value{
`TRUE` if `adduct` is valid and `FALSE` otherwise.
}
\description{
Check whether an adduct string matches the package's supported syntax.
}
\examples{
check_adduct_formula("(M+H)+") # TRUE
check_adduct_formula("(M+X)+") # FALSE

}
