% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_check_functions.R
\name{check_chemical_element}
\alias{check_chemical_element}
\title{Validate a Chemical Element Symbol}
\usage{
check_chemical_element(element)
}
\arguments{
\item{element}{A character string to validate.}
}
\value{
`TRUE` if `element` is a valid chemical element symbol and `FALSE`
  otherwise.
}
\description{
Check whether a character string is a valid chemical element symbol.
}
\examples{
check_chemical_element("Fe") # TRUE
check_chemical_element("Fz") # FALSE

}
