% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_check_functions.R
\name{check_chemical_formula}
\alias{check_chemical_formula}
\title{Validate a Chemical Formula}
\usage{
check_chemical_formula(formula)
}
\arguments{
\item{formula}{A character string to validate.}
}
\value{
`TRUE` if `formula` is valid and `FALSE` otherwise.
}
\description{
Check whether a chemical formula follows the package's validation rules.
}
\examples{
check_chemical_formula("H2O") # TRUE
check_chemical_formula("2H") # FALSE

}
