% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_conversion.R
\name{check_spectrum_type}
\alias{check_spectrum_type}
\title{Check the Type of a Spectrum-Like Object}
\usage{
check_spectrum_type(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Character string: "Spectra", "data.frame", "matrix", or "unknown"
}
\description{
Internal helper that classifies supported spectrum input types.
}
\keyword{internal}
