% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classyfire.R
\docType{class}
\name{classyfire}
\alias{classyfire}
\title{S4 Class for ClassyFire Results}
\description{
Store the result of a ClassyFire query, including compound information,
taxonomy, free-text description, and external descriptors.
}
\section{Slots}{

\describe{
\item{\code{compound_info}}{A tibble containing basic compound information such as
SMILES, InChIKey, Formula, and Mass.}

\item{\code{taxonomy_tree}}{A tibble representing the taxonomic tree of the compound
that includes classifications like Kingdom, Superclass, Class, and Subclass.}

\item{\code{classification_info}}{A tibble containing a detailed taxonomic
classification of the compound.}

\item{\code{description}}{A tibble holding the compound's description from ClassyFire.}

\item{\code{external_descriptors}}{A tibble with external descriptors related to the
compound.}
}}

\seealso{
\code{\link[methods:show]{show}}
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
