% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_formula_adduct.R
\name{combine_formula_adduct}
\alias{combine_formula_adduct}
\title{Sum a Molecular Formula with an Adduct}
\usage{
combine_formula_adduct(formula = "C9H11NO2", adduct = "M-H2O+H")
}
\arguments{
\item{formula}{A character string containing the base molecular formula.}

\item{adduct}{A character string containing the adduct expression, such as
`"M-H2O+H"` or `"M+CH3COOH"`.}
}
\value{
A character string containing the summed formula, or `NA` if the
  result is invalid.
}
\description{
Combine a molecular formula with an adduct expression and return the
resulting summed formula.
}
\examples{
combine_formula_adduct(formula = "C9H11NO2", adduct = "M+H")
combine_formula_adduct(formula = "C9H11NO2", adduct = "M+")
combine_formula_adduct(formula = "C9H11NO2", adduct = "M+CH3COOH")
combine_formula_adduct(formula = "C9H11", adduct = "M-H20")
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
