% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MF_formatting.R
\name{consolidateMF}
\alias{consolidateMF}
\title{consolidateMF}
\usage{
consolidateMF(x)
}
\arguments{
\item{x}{element count vector}
}
\description{
Merges duplicates in an element count vector
}
