% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MF_formatting.R
\name{consolidateSF}
\alias{consolidateSF}
\title{'consolidateSF}
\usage{
consolidateSF(x)
}
\arguments{
\item{x}{element count vector}
}
\description{
Merges duplicates in an element count vector
}
