% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_chemical_formula_funcitons.R
\name{convert_precursor_mz2accurate_mass}
\alias{convert_precursor_mz2accurate_mass}
\title{Convert Precursor m/z to Accurate Mass}
\usage{
convert_precursor_mz2accurate_mass(
  precursor_mz = 805.559,
  adduct = "(2M-2H+Na)-"
)
}
\arguments{
\item{precursor_mz}{Numeric precursor m/z value.}

\item{adduct}{Character string describing the adduct, for example
`"(M-H)-"` or `"(2M-2H+Na)-"`.}
}
\value{
A numeric value giving the calculated neutral mass.
}
\description{
Calculate neutral mass from a precursor m/z value and adduct.
}
\examples{
convert_precursor_mz2accurate_mass(805.559, "(2M-2H+Na)-")
convert_precursor_mz2accurate_mass(805.559, "(M-H)-")
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
