% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peptides.R
\name{countAAs}
\alias{countAAs}
\title{countAAs}
\usage{
countAAs(seq)
}
\arguments{
\item{seq}{one or more peptide sequences in a character vector}
}
\value{
a list of named numeric vectors
}
\description{
count amino acids in a sequence
}
