% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_names.R
\name{deduplicate_names}
\alias{deduplicate_names}
\title{Make Duplicated Names Unique}
\usage{
deduplicate_names(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector where duplicated names are made unique by
  appending a sequence number.
}
\description{
Append sequence numbers to duplicated character values.
}
\examples{
vec <- c("apple", "orange", "apple", "banana", "orange")
deduplicate_names(vec)

}
