% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_os.R
\name{detect_os}
\alias{detect_os}
\title{Determine the Operating System of the Current R Session}
\usage{
detect_os()
}
\value{
A character scalar such as `"windows"`, `"osx"`, or `"linux"`.
}
\description{
This function identifies the operating system for the current R session and
returns it in a standardized format.
}
\examples{
detect_os()
}
