% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Spectra.R
\name{findPatterns}
\alias{findPatterns}
\title{findPatterns}
\usage{
findPatterns(speclist, patterns, ppm = 5, mzdiff = 0)
}
\arguments{
\item{speclist}{list of spectra (matrices, all with two columns: \code{mz} and \code{intensity})}

\item{patterns}{a named list of patterns (numeric vectors of mz values to look for)}

\item{ppm}{min difference between peaks in ppm}

\item{mzdiff}{min difference between peaks in m/z}
}
\value{
a list (same length as speclist) of logical vectors specifying if patterns are matched
}
\description{
find mz patterns in a list of spectra
}
