% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_get_compound_class.R
\name{get_compound_class}
\alias{get_compound_class}
\title{Retrieve Compound Classification Information}
\usage{
get_compound_class(
  inchikey = "QZDWODWEESGPLC-UHFFFAOYSA-N",
  server = "http://classyfire.wishartlab.com/entities/",
  sleep = 5
)
}
\arguments{
\item{inchikey}{A character string representing the InChIKey of the compound.
Default is "QZDWODWEESGPLC-UHFFFAOYSA-N".}

\item{server}{A character string representing the base URL of the ClassyFire server.
Default is "http://classyfire.wishartlab.com/entities/".}

\item{sleep}{A numeric value indicating the time (in seconds) to pause between the
request and fetching the result to avoid overloading the server. Default is 5 seconds.}
}
\value{
An S4 object of class `"classyfire"`, or `NA` on failure.
}
\description{
`get_compound_class()` is deprecated. Please use
`retrieve_compound_class()` instead.
}
\examples{
\dontrun{
get_compound_class(
  inchikey = "QZDWODWEESGPLC-UHFFFAOYSA-N",
  server = "http://classyfire.wishartlab.com/entities/",
  sleep = 5
)
}
}
