% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_get_dp.R
\name{get_dp}
\alias{get_dp}
\title{Compute a Weighted Dot Product Between Two Intensity Vectors}
\usage{
get_dp(exp.int, lib.int)
}
\arguments{
\item{exp.int}{A numeric vector representing the intensity values of the experimental spectrum.}

\item{lib.int}{A numeric vector representing the intensity values of the library spectrum.}
}
\value{
A numeric value representing the weighted dot product.
}
\description{
`get_dp()` is deprecated. Please use `calculate_dot_product()` instead.
}
\examples{
exp.int <- c(0.10, 0.35, 1.00, 0.45)
lib.int <- c(0.12, 0.30, 0.95, 0.50)
get_dp(exp.int = exp.int, lib.int = lib.int)
}
