% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_get_os.R
\name{get_os}
\alias{get_os}
\title{Determine the Operating System of the Current R Session}
\usage{
get_os()
}
\value{
A character scalar such as `"windows"`, `"osx"`, or `"linux"`.
}
\description{
`get_os()` is deprecated. Please use `detect_os()` instead.
}
\examples{
get_os()
}
