% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_getDP.R
\name{getDP}
\alias{getDP}
\title{Compute a Dot Product Between Two Intensity Vectors}
\usage{
getDP(exp.int, lib.int)
}
\arguments{
\item{exp.int}{A numeric vector representing the intensity values of the experimental spectrum.}

\item{lib.int}{A numeric vector representing the intensity values of the library spectrum.}
}
\value{
A numeric value representing the weighted dot product between
  `exp.int` and `lib.int`.
}
\description{
`getDP()` is deprecated. Please use `calculate_dot_product()` instead.
}
\examples{
exp.int <- c(0.10, 0.35, 1.00, 0.45)
lib.int <- c(0.12, 0.30, 0.95, 0.50)

getDP(exp.int = exp.int, lib.int = lib.int)
}
