% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{getExactMass.character}
\alias{getExactMass.character}
\title{getExactMass.character}
\usage{
\method{getExactMass}{character}(x, ...)
}
\arguments{
\item{x}{molecular formula as a character()}

\item{...}{additional arguments to \code{\link[Rdisop]{getMolecule}}}
}
\description{
getExactMass.character
}
