% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{getExactMass.MFobject}
\alias{getExactMass.MFobject}
\title{getExactMass.MFobject}
\usage{
\method{getExactMass}{MFobject}(x, ...)
}
\arguments{
\item{x}{molecular formula as an MFobject}

\item{...}{additional arguments to \code{\link[Rdisop]{getMolecule}}}
}
\description{
getExactMass.MFobject
}
