% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{getExactMass}
\alias{getExactMass}
\title{getExactMass}
\usage{
getExactMass(x, ...)
}
\arguments{
\item{x}{molecular formula as a character() or MFobject}
}
\description{
get the exact mass of a molecule
}
