% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{getSenior3.MFobject}
\alias{getSenior3.MFobject}
\title{getSenior3}
\usage{
getSenior3.MFobject(MF)
}
\arguments{
\item{MF}{MFobject as constructed by \code{\link{makeMF}}}
}
\description{
calculate the sum of valences minus twice (the number of atoms minus 1) (to check for Senior's third theorem, "Golden Rule #2")[1][2][3]
}
\references{
\enumerate{
\item Kind T, Fiehn O (2007) Seven Golden Rules for heuristic filtering of molecular formulas obtained by accurate mass spectrometry. BMC Bioinformatics 8:105. doi: 10.1186/1471-2105-8-105
\item Senior JK (1951) Partitions and Their Representative Graphs. Am J Math 73:663. doi: 10.2307/2372318
\item Morikawa T, Newbold BT (2003) Analogous odd-even parities in mathematics and chemistry. Chemistry 12:445–450
}
}
