% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MF_prediction.R
\name{getSenior3}
\alias{getSenior3}
\title{getSenior3}
\usage{
getSenior3(MF, vals = c(C = 4, H = 1, Br = 1, Cl = 1, F = 1, N = 5, O =
  2, P = 5, S = 6, Si = 4, Na = 1, K = 1, Mg = 2, Ca = 2))
}
\arguments{
\item{MF}{MFobject as generated by \code{\link{makeMF}}}

\item{vals}{maximum valences for a variety of elements}
}
\description{
calculate the sum of valences minus twice (the number of atoms minus 1) (to check for Senior's third theorem, "Golden Rule #2")[1][2][3]
}
\references{
\enumerate{
\item Kind T, Fiehn O (2007) Seven Golden Rules for heuristic filtering of molecular formulas obtained by accurate mass spectrometry. BMC Bioinformatics 8:105. doi: 10.1186/1471-2105-8-105
\item Senior JK (1951) Partitions and Their Representative Graphs. Am J Math 73:663. doi: 10.2307/2372318
\item Morikawa T, Newbold BT (2003) Analogous odd-even parities in mathematics and chemistry. Chemistry 12:445–450
}
}
