% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_getSpectraMatchScore.R
\name{getSpectraMatchScore}
\alias{getSpectraMatchScore}
\title{Compute a Similarity Score Between Two Mass Spectra}
\usage{
getSpectraMatchScore(
  exp.spectrum,
  lib.spectrum,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  fraction.weight = 0.2,
  dp.forward.weight = 0.7,
  dp.reverse.weight = 0.1,
  remove_fragment_intensity_cutoff = 0
)
}
\arguments{
\item{exp.spectrum}{The experimental spectrum. Can be:
\itemize{
  \item A data frame with 'mz' and 'intensity' columns
  \item A Spectra object from the Spectra package
}}

\item{lib.spectrum}{The library/reference spectrum. Can be:
\itemize{
  \item A data frame with 'mz' and 'intensity' columns
  \item A Spectra object from the Spectra package
}}

\item{ppm.tol}{A numeric value indicating the ppm tolerance.}

\item{mz.ppm.thr}{A numeric value for m/z threshold for ppm calculation.}

\item{fraction.weight}{A numeric value for weight of fraction component in score.}

\item{dp.forward.weight}{A numeric value for weight of forward dot product component in score.}

\item{dp.reverse.weight}{A numeric value for weight of reverse dot product component in score.}

\item{remove_fragment_intensity_cutoff}{A numeric value specifying the intensity cutoff for filtering out fragments in the spectra. Default is 0, which means no fragments are filtered out based on intensity.}
}
\value{
A numeric value representing the similarity score between
  `exp.spectrum` and `lib.spectrum`.
}
\description{
`getSpectraMatchScore()` is deprecated. Please use
`calculate_spectra_match_score()` instead.
}
\examples{
exp.spectrum <- data.frame(
  mz = c(100.00, 124.04, 150.06, 180.08, 205.10),
  intensity = c(10, 35, 100, 45, 20)
)
lib.spectrum <- data.frame(
  mz = c(100.01, 124.05, 150.06, 180.09, 220.12),
  intensity = c(12, 30, 95, 50, 10)
)

getSpectraMatchScore(exp.spectrum, lib.spectrum, ppm.tol = 50)
}
