% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{getTopIsotope}
\alias{getTopIsotope}
\title{getTopIsotope}
\usage{
getTopIsotope(x, ...)
}
\arguments{
\item{x}{molecular formula as a character() or MFobject}

\item{...}{additional arguments to \code{\link[Rdisop]{getMolecule}}}
}
\description{
get the mass of the most abundant isotope for a molecular formula
}
\details{
Warning: if mixing positive and negative values in one molecular formula,
 this may not give the correct result
}
