% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{install_fastgit}
\alias{install_fastgit}
\title{install_fastgit}
\usage{
install_fastgit(pkg, from = c("gitee", "gitlab", "github"), ...)
}
\arguments{
\item{pkg}{pkg name from github, gitlab or gitee, "name/repo" format}

\item{from}{gitlab, github or gitee.}

\item{...}{Other parameters for install_git}
}
\value{
`NULL`. This function is called for its side effect of installing packages from specified Git repository mirrors.
}
\description{
install packages from fastgit. Credit to Shixiang Wang
}
\examples{
\dontrun{
# Install a package from GitHub using the fastgit mirror
install_fastgit(pkg = "tidyverse/ggplot2", from = "github")

# Install a package from Gitee
install_fastgit(pkg = "yourusername/yourpkg", from = "gitee")
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
