% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-chemical_formula_funcitons.R
\name{is_valid_chemical_formula}
\alias{is_valid_chemical_formula}
\title{Check Validity of a Chemical Formula}
\usage{
is_valid_chemical_formula(formula)
}
\arguments{
\item{formula}{Character string representing the chemical formula to be checked.}
}
\value{
Logical. Returns `TRUE` if the formula is valid and `FALSE` otherwise.
}
\description{
A function to validate chemical formulas based on the list of recognized chemical elements as of 2022.
}
\details{
Check Validity of a Chemical Formula

Determines the validity of a chemical formula based on the known chemical elements up to 2022.
The function checks whether all the symbols in the formula belong to the list of known chemical elements.
It doesn't consider charges or other non-element symbols in the formula.
}
\examples{
is_valid_chemical_formula("H2O")      # TRUE
is_valid_chemical_formula("C6H12O6")  # TRUE
is_valid_chemical_formula("H2ZO4")    # FALSE

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
