% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_masstools_packages.R
\name{list_masstools_packages}
\alias{list_masstools_packages}
\title{List Packages Imported by masstools}
\usage{
list_masstools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical. If `TRUE`, include `"masstools"` in the
returned character vector.}
}
\value{
A character vector of package names listed in the package Imports.
}
\description{
List Packages Imported by masstools
}
\examples{
list_masstools_packages()
}
