% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_conversion.R
\name{list_to_spectra}
\alias{list_to_spectra}
\title{Convert a Legacy Spectrum List to a Spectra Object}
\usage{
list_to_spectra(spectrum_list)
}
\arguments{
\item{spectrum_list}{A list where each element contains 'info' and 'spec'}
}
\value{
A Spectra object
}
\description{
Converts the legacy list format (from read_mgf/read_mzxml) to a Spectra object.
}
\keyword{internal}
