% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MF_formatting.R
\name{makeMF}
\alias{makeMF}
\title{makeMF}
\usage{
makeMF(s, forcelist = F)
}
\arguments{
\item{s}{Molecula formula string}

\item{forcelist}{if \code{TRUE}, will always return a list, even when s is of length 1.}
}
\description{
construct an MFobject from a molecula formula string
}
