% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masstools_logo.R
\name{masstools_logo}
\alias{masstools_logo}
\title{Print the masstools Startup Logo}
\usage{
masstools_logo()
}
\value{
This function is called for its side effects and returns `NULL`
  invisibly.
}
\description{
Print the masstools Startup Logo
}
\examples{
masstools_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
