% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_masstools_packages.R
\name{masstools_packages}
\alias{masstools_packages}
\title{List Packages Imported by masstools}
\usage{
masstools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical. If `TRUE`, include `"masstools"` in the
returned character vector.}
}
\value{
A character vector of package names listed in the package Imports.
}
\description{
`masstools_packages()` is deprecated. Please use
`list_masstools_packages()` instead.
}
\examples{
masstools_packages()
}
