% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_utilities.R
\name{matchNumbers}
\alias{matchNumbers}
\title{matchNumbers}
\usage{
matchNumbers(v1, v2, ppm = 5, abs = 0.001)
}
\arguments{
\item{v1, v2}{numeric vectors}

\item{ppm}{matching tolerance in ppm}

\item{abs}{matching tolerance in absolute numbers}
}
\value{
an indexing matrix of matching values, 
with matching pairs in each row, column 1 is the v1 match,
 row 2 is the v2 match.
}
\description{
match numbers between two numeric vectors with tolerances. 
A match is defined by complying wit either the ppm OR the 
abs tolerance values.
}
