% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_ms2Match.R
\name{ms2Match}
\alias{ms2Match}
\title{Match Fragments Between Two Spectra}
\usage{
ms2Match(exp.spectrum, lib.spectrum, ppm.tol = 30, mz.ppm.thr = 400)
}
\arguments{
\item{exp.spectrum}{The experimental spectrum. Can be:
\itemize{
  \item A data frame with columns for m/z and intensity
  \item A Spectra object from the Spectra package
}}

\item{lib.spectrum}{The library/reference spectrum. Can be:
\itemize{
  \item A data frame with columns for m/z and intensity
  \item A Spectra object from the Spectra package
}}

\item{ppm.tol}{A numeric value specifying the ppm tolerance for matching fragments. Default is 30.}

\item{mz.ppm.thr}{A numeric value for the m/z threshold. Default is 400.}
}
\value{
A data frame with columns for library index, experimental index,
  library m/z, library intensity, experimental m/z, and experimental
  intensity.
}
\description{
`ms2Match()` is deprecated. Please use `ms2_match()` instead.
}
\examples{
exp.spectrum <- data.frame(
  mz = c(100.0000, 124.0400, 150.0600, 180.0800, 205.1000),
  intensity = c(10, 35, 100, 45, 20)
)
lib.spectrum <- data.frame(
  mz = c(100.0015, 124.0395, 150.0610, 180.0820, 220.1200),
  intensity = c(12, 30, 95, 50, 10)
)

ms2Match(exp.spectrum, lib.spectrum, ppm.tol = 50)
}
