% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_name_duplicated.R
\name{name_duplicated}
\alias{name_duplicated}
\title{Make Duplicated Names Unique}
\usage{
name_duplicated(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector where duplicated names are made unique by
  appending a sequence number.
}
\description{
`name_duplicated()` is deprecated. Please use `deduplicate_names()`
instead.
}
\examples{
vec <- c("apple", "orange", "apple", "banana", "orange")
name_duplicated(vec)

}
