% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{Ops.MFobject}
\alias{Ops.MFobject}
\title{Ops.MFobject}
\usage{
\method{Ops}{MFobject}(a, b)
}
\arguments{
\item{a}{numeric or MFobject}

\item{b}{numeric or MFobject}
}
\description{
Operator behavior for MFobject S3 class
}
