% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_chemical_formula_funcitons.R
\name{parse_chemical_formula}
\alias{parse_chemical_formula}
\title{Parse a Molecular Formula}
\usage{
parse_chemical_formula(formula)
}
\arguments{
\item{formula}{Character string containing a molecular formula such as
`"H2O"` or `"C6H12O6"`.}
}
\value{
A data frame with columns `Element` and `Count`.
}
\description{
Parse a molecular formula into element symbols and counts.
}
\details{
Parse a Molecular Formula

Break a molecular formula into element symbols and counts.
}
\examples{
parse_chemical_formula("H2O")       # Returns a data frame with H:2 and O:1
parse_chemical_formula("C6H12O6")   # Returns a data frame with C:6, H:12, and O:6
parse_chemical_formula("Fe2O3")     # Returns a data frame with Fe:2 and O:3

}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
