% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_utilities.R
\name{permutateMass}
\alias{permutateMass}
\title{permutateMass}
\usage{
permutateMass(baseMass, modifications, sep = "\\\\|",
  unmodifiedTag = "")
}
\arguments{
\item{baseMass}{numeric mass value to permutate}

\item{modifications}{data.frame with columns name, formula, min, max.}

\item{sep}{separator between individual modifications in output;
has to be escaped with \\ if it is a special character for regex}

\item{unmodifiedTag}{which tag to use for non-modified items. If NULL,
will preceed modification tag with 0 (e.g. 0Me)}
}
\value{
a data.fram, the mass in the resulting data.frame is the  monoisotopic mass.
}
\description{
Use a data.frame with molecular formulas and combine them
}
\details{
More information
\subsection{ modifications data.frame}{
Columns that need to be in the modifications data.frame:
\itemize{
\item{min} minimum number of times this modification can be applied
\item{max} maximum number of times this modification can be applied
\item{mass} mass of this modification
\item{tag} label for this modification (will automatically be preceded by 
number of these modifications applied)
}
columns min and max should be the minimum and maximum times a modification 
is allowed in the combinations, can be positive, negative, or 0.
}
}
\seealso{
\code{\link{permutatePeptideMass}}
}
