% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peptides.R
\name{permutatePeptideMass}
\alias{permutatePeptideMass}
\title{permutatePeptideMass}
\usage{
permutatePeptideMass(df, modifications, sequenceCol = "seq",
  massCol = "mz", chargeCol = NULL)
}
\arguments{
\item{df}{data.frame with at least two columns, one with a non-empty 
sequence string and one with a mass for that sequence}

\item{modifications}{a data.frame with colums \code{AAs}}

\item{sequenceCol}{name of the sequence column in df}

\item{massCol}{name of the mass column in df}

\item{chargeCol}{name of column that specifies charge state (used to adjust 
modification masses)}
}
\description{
Apply a set of mass modifications to a peptide with a known mass
}
\details{
\subsection{modifications}{
Columns that need to be in the \code{modifications data.frame}:
\describe{
\item{AAs}{containing letters for the amino acids (e.g. "TS")
affected by this modification}
\item{min}{minimum number of times this modification can be applied}
\item{max}{maximum number of times this modification can be applied}
\item{mass}{mass of this modification}
\item{tag}{ label for this modification (will automatically be preceded by 
number of these modifications applied)}
}
min and max can be positive, negative or 0. The larger *absolute* value has 
to be in the \code{max} column
}
}
\examples{
peptides = data.frame(seq = "ASDFGHKLIPYTREWQCNM",mz = 2295.056)
mods = data.frame(AAs = c("TS", ""),
 min = c(0,1), max = c(-10,2),
  mass = c(18.01,12.0), tag = c("DH","C") )

permutatePeptideMass(peptides, mods)


}
\seealso{
\code{\link{permutateMass}}
}
