% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ms2.R
\name{plot_ms2}
\alias{plot_ms2}
\alias{plot_ms2.data.frame}
\alias{plot_ms2.matrix}
\title{Plot MS2 Spectra Comparisons}
\usage{
plot_ms2(
  spectrum1,
  spectrum2,
  spectrum1_name = "spectrum1",
  spectrum2_name = "spectrum2",
  range.mz,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity",
  col1 = "red",
  col2 = "black",
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  legend.title.size = 15,
  legend.text.size = 15,
  interactive_plot = FALSE
)

\method{plot_ms2}{data.frame}(
  spectrum1,
  spectrum2,
  spectrum1_name = "spectrum1",
  spectrum2_name = "spectrum2",
  range.mz,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity",
  col1 = "red",
  col2 = "black",
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  legend.title.size = 15,
  legend.text.size = 15,
  interactive_plot = FALSE
)

\method{plot_ms2}{matrix}(
  spectrum1,
  spectrum2,
  spectrum1_name = "spectrum1",
  spectrum2_name = "spectrum2",
  range.mz,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity",
  col1 = "red",
  col2 = "black",
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  legend.title.size = 15,
  legend.text.size = 15,
  interactive_plot = FALSE
)
}
\arguments{
\item{spectrum1}{A spectrum object representing the first MS2 spectrum.}

\item{spectrum2}{A spectrum object representing the second MS2 spectrum.}

\item{spectrum1_name}{A character string specifying the name/label for the first spectrum. Default is "spectrum1".}

\item{spectrum2_name}{A character string specifying the name/label for the second spectrum. Default is "spectrum2".}

\item{range.mz}{Numeric vector of length 2 specifying the range of m/z values to display in the plot.}

\item{ppm.tol}{Numeric value specifying the ppm tolerance for matching peaks between the spectra. Default is 30.}

\item{mz.ppm.thr}{Numeric value specifying the threshold for m/z ppm difference. Default is 400.}

\item{xlab}{Character string specifying the x-axis label. Default is "Mass to charge ratio (m/z)".}

\item{ylab}{Character string specifying the y-axis label. Default is "Relative intensity".}

\item{col1}{Character string specifying the color for the first spectrum. Default is "red".}

\item{col2}{Character string specifying the color for the second spectrum. Default is "black".}

\item{title.size}{Numeric value specifying the font size for the plot title. Default is 15.}

\item{lab.size}{Numeric value specifying the font size for x and y labels. Default is 15.}

\item{axis.text.size}{Numeric value specifying the font size for axis text. Default is 15.}

\item{legend.title.size}{Numeric value specifying the font size for the legend title. Default is 15.}

\item{legend.text.size}{Numeric value specifying the font size for the legend text. Default is 15.}

\item{interactive_plot}{Logical indicating whether the plot should be interactive (e.g., using plotly). Default is FALSE.}
}
\value{
A plot visualizing the comparison between the two provided MS2 spectra.
}
\description{
Visualize one or two MS2 spectra, with matched peaks shown in a mirrored
comparison when two spectra are provided.
}
\examples{
spectrum1 <- data.frame(
  mz = c(100.00, 124.04, 150.06, 180.08, 205.10),
  intensity = c(10, 35, 100, 45, 20)
)
spectrum2 <- data.frame(
  mz = c(100.01, 124.05, 150.06, 180.09, 220.12),
  intensity = c(12, 30, 95, 50, 10)
)

plot_ms2(spectrum1, spectrum2)

spectrum1 <- data.frame(
  mz = c(100.00, 124.04, 150.06, 180.08, 205.10),
  intensity = c(10, 35, 100, 45, 20)
)
spectrum2 <- data.frame(
  mz = c(100.01, 124.05, 150.06, 180.09, 220.12),
  intensity = c(12, 30, 95, 50, 10)
)

plot_ms2(spectrum1, spectrum2)
plot_ms2(spectrum1)

\dontrun{
plot_ms2(spectrum1, spectrum2, interactive_plot = TRUE)
}
spectrum1_mat <- matrix(
  c(100.00, 10,
    124.04, 35,
    150.06, 100,
    180.08, 45,
    205.10, 20),
  ncol = 2,
  byrow = TRUE
)
spectrum2_mat <- matrix(
  c(100.01, 12,
    124.05, 30,
    150.06, 95,
    180.09, 50,
    220.12, 10),
  ncol = 2,
  byrow = TRUE
)

plot_ms2(spectrum1_mat, spectrum2_mat)

}
\seealso{
The specific plotting methods associated with different spectrum classes that implement this generic function.
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
