% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peptides.R
\name{plotAnnotatedPeptide}
\alias{plotAnnotatedPeptide}
\title{plotAnnotatedPeptide}
\usage{
plotAnnotatedPeptide(peaklabels, sequence, parseLabels = F,
  sortby = "intensitySpec", cx = 2, yoffset = 0)
}
\arguments{
\item{peaklabels}{data.frame with labeling informationi,
as returned by \code{\link{annotateSpectrum}}}

\item{sequence}{peptide sequence (character)}

\item{parseLabels}{optionally apply parse() to
peaklabels$label and add as label (NYI)}

\item{sortby}{sort (decreasing) by this column. Only the top 2 ions 
for each position (for both a,b,c ions AND x,y,z ions) will be shown}

\item{cx}{font size factor for the sequence characters,
other font sizes will be scaled accordingly}

\item{yoffset}{y axis offset}
}
\description{
plot an annotated peptide sequence
}
