% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{print.MFobject}
\alias{print.MFobject}
\title{print.MFobject}
\usage{
\method{print}{MFobject}(x, ...)
}
\arguments{
\item{x}{MFobject}

\item{...}{other arguments to \code{base::print()}}
}
\description{
Print method for MFobject S3 class
}
