% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_read_mgf4database.R
\name{read_mgf4database}
\alias{read_mgf4database}
\title{Read MGF Files for Database Workflows}
\usage{
read_mgf4database(file)
}
\arguments{
\item{file}{A character vector of MGF file paths.}
}
\value{
A list of spectra in the database workflow format.
}
\description{
`read_mgf4database()` is deprecated. Please use `read_mgf_database()`
instead.
}
\examples{
file_path <- system.file("extdata", "example.mgf", package = "masstools")
result <- read_mgf4database(file_path)
length(result)
}
