% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MF_formatting.R
\name{remakeMF}
\alias{remakeMF}
\title{remakeMF}
\usage{
remakeMF(x)
}
\arguments{
\item{x}{Molecula formula string}
}
\description{
Converts an element count vector back into a molecula formula string
}
