% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_functions.R
\name{remove_noise}
\alias{remove_noise}
\title{Remove Noisy Fragments from a Spectrum}
\usage{
remove_noise(spec, ppm.ms2match = 30, mz.ppm.thr = 400)
}
\arguments{
\item{spec}{A mass spectrum. Can be:
\itemize{
  \item A data frame with columns for m/z and intensity
  \item A Spectra object from the Spectra package
}}

\item{ppm.ms2match}{A numeric value specifying the ppm threshold for identifying close fragments. Default is 30.}

\item{mz.ppm.thr}{A numeric value for the m/z threshold below which the m/z value will be set to this threshold. Default is 400.}
}
\value{
A spectrum in the same format as input:
  \itemize{
    \item If input is data.frame: returns data.frame
    \item If input is Spectra: returns Spectra object
  }
}
\description{
Remove fragments with nearly identical m/z values and keep the more intense
peak.
}
\details{
Peaks are ordered by m/z, adjacent peaks are compared using the ppm
threshold, and the less intense peak is removed when two peaks are too
close. The function accepts both data frames and `Spectra` objects.
}
\examples{
spectrum <- data.frame(
  mz = c(100.0000, 100.0010, 124.0400, 150.0600, 150.0610),
  intensity = c(10, 5, 35, 100, 20)
)

remove_noise(spectrum, ppm.ms2match = 30)

# The function also accepts Spectra objects.
\dontrun{
spec_obj <- df_to_spectra(spectrum)

remove_noise(spec_obj, ppm.ms2match = 30)
}
}
