% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_removeNoise.R
\name{removeNoise}
\alias{removeNoise}
\title{Remove Noisy Fragments from a Spectrum}
\usage{
removeNoise(spec, ppm.ms2match = 30, mz.ppm.thr = 400)
}
\arguments{
\item{spec}{A mass spectrum. Can be:
\itemize{
  \item A data frame with columns for m/z and intensity
  \item A Spectra object from the Spectra package
}}

\item{ppm.ms2match}{A numeric value specifying the ppm threshold for identifying close fragments. Default is 30.}

\item{mz.ppm.thr}{A numeric value for the m/z threshold below which the m/z value will be set to this threshold. Default is 400.}
}
\value{
A spectrum in the same format as input.
}
\description{
`removeNoise()` is deprecated. Please use `remove_noise()` instead.
}
\examples{
spectrum <- data.frame(
  mz = c(100.0000, 100.0010, 124.0400, 150.0600, 150.0610),
  intensity = c(10, 5, 35, 100, 20)
)

removeNoise(spectrum, ppm.ms2match = 30)
}
