% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_request_metabolite_id_systems.R
\name{request_metabolite_id_systems}
\alias{request_metabolite_id_systems}
\title{Retrieve Supported Identifier Conversion Systems}
\usage{
request_metabolite_id_systems(
  server = c("cts.fiehnlab", "chemspider", "openai"),
  source_format = c("local", "online")
)
}
\arguments{
\item{server}{Character. The server to retrieve supported identifier formats from.}

\item{source_format}{Character. Specifies whether to use a **local** or **online** source.}
}
\value{
A data frame or vector containing the supported identifier
  conversions for the selected backend.
}
\description{
`request_metabolite_id_systems()` is deprecated. Please use
`list_metabolite_id_systems()` instead.
}
\examples{
request_metabolite_id_systems(server = "cts.fiehnlab", source_format = "local")
}
