% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_project.R
\name{setwd_project}
\alias{setwd_project}
\title{setwd_project}
\usage{
setwd_project()
}
\value{
nothing
}
\description{
Set work directory where Rproj object in.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
