% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_conversion.R
\name{spectra_to_df}
\alias{spectra_to_df}
\title{Convert a Spectra Object to a Data Frame}
\usage{
spectra_to_df(spectra, index = 1)
}
\arguments{
\item{spectra}{A `Spectra` object.}

\item{index}{Integer. Which spectrum to extract.}
}
\value{
A data frame with `mz` and `intensity` columns.
}
\description{
Extract one spectrum from a `Spectra` object and return it as a data frame.
}
\examples{
spectrum_df <- data.frame(
  mz = c(100.00, 124.04, 150.06),
  intensity = c(10, 35, 100)
)
spec <- df_to_spectra(spectrum_df)

spectra_to_df(spec)
}
