% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_conversion.R
\name{spectra_to_list}
\alias{spectra_to_list}
\title{Convert Multiple Spectra to a List Representation}
\usage{
spectra_to_list(spectra)
}
\arguments{
\item{spectra}{A Spectra object}
}
\value{
A list where each element contains 'info' and 'spec'
}
\description{
Converts a Spectra object with multiple spectra to the legacy list format
used by read_mgf and read_mzxml functions.
}
\keyword{internal}
