% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_split_formula.R
\name{split_formula}
\alias{split_formula}
\title{Parse a Molecular Formula}
\usage{
split_formula(formula = "C9H11NO2")
}
\arguments{
\item{formula}{Character string containing a molecular formula.}
}
\value{
A data frame with columns `element` and `number`.
}
\description{
`split_formula()` is deprecated. Please use `parse_chemical_formula()`
instead.
}
\examples{
split_formula(formula = "C9H11NO2")
split_formula(formula = "CH3")
split_formula(formula = "CH3C")
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
