% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_sum_formula.R
\name{sum_formula}
\alias{sum_formula}
\title{Sum a Molecular Formula with an Adduct}
\usage{
sum_formula(formula = "C9H11NO2", adduct = "M-H2O+H")
}
\arguments{
\item{formula}{A character string containing the base molecular formula.}

\item{adduct}{A character string containing the adduct expression, such as
`"M-H2O+H"` or `"M+CH3COOH"`.}
}
\value{
A character string containing the summed formula, or `NA` if the
  result is invalid.
}
\description{
`sum_formula()` is deprecated. Please use `combine_formula_adduct()`
instead.
}
\examples{
sum_formula(formula = "C9H11NO2", adduct = "M+H")
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
