% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_deprecated_functions.R
\name{trans_id_database}
\alias{trans_id_database}
\title{Retrieve Supported Databases for Chemical Identifier Conversion}
\usage{
trans_id_database(server = c("cts.fiehnlab", "chemspider"))
}
\arguments{
\item{server}{A character string indicating which server to use. Possible values are "cts.fiehnlab" or "chemspider".}
}
\value{
A list of supported databases. If using the `cts.fiehnlab` server, the list will contain two data frames: 'From' and 'To'. If using the `chemspider` server, the result will be a single data frame with 'From' and 'To' columns.
}
\description{
This function lists the databases supported for chemical identifier conversion by either the `cts.fiehnlab` or `chemspider` services.
}
\note{
This function is deprecated.
}
\examples{
\dontrun{
trans_id_database(server = "cts.fiehnlab")
}
}
