% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Spectra.R
\name{annotateSpectrum}
\alias{annotateSpectrum}
\title{annotateSpectrum}
\usage{
annotateSpectrum(labels, spectrum, mzlabel = F, unmodifiedTag = "",
  ppm = 5, abs = 0)
}
\arguments{
\item{labels}{a data.frame as returned by permutatePeptideMass}

\item{spectrum}{a mass spectrum as data.frame or matrix}

\item{mzlabel}{if TRUE, will add mz values to the label}

\item{unmodifiedTag}{what tag to use for unmodified fragment ions}

\item{ppm}{relative matching tolerance in ppm}

\item{abs}{matching tolerance in absolute numbers}
}
\value{
plots an annotated peptide sequence in the current plotting device
}
\description{
Match peaks in a spectrum with those in a labeling data.frame 
and prepare them for plotting functions. Matches have to be within abs OR ppm
}
