% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_chemical_formula_funcitons.R
\name{calculate_mass}
\alias{calculate_mass}
\title{Calculate Mass of a Chemical Formula}
\usage{
calculate_mass(formula, which = c("exact_mass", "average_mass"))
}
\arguments{
\item{formula}{Character string representing the chemical formula for which the mass is to be computed.}

\item{which}{Character string indicating the type of mass to be computed. Possible values are "exact_mass" or "average_mass". Default is "exact_mass".}
}
\value{
A numeric value representing the computed mass of the given chemical formula based on the specified type (exact or average).
}
\description{
Computes the mass of a given chemical formula based either on the exact mass or the average mass of its constituent elements.
}
\details{
Calculate Mass of a Chemical Formula

This function computes the mass of a given chemical formula based on either exact or average mass of its constituent elements.
}
\examples{
calculate_mass("H2O")                      # Using default exact_mass
calculate_mass("H2O", which = "average_mass")
calculate_mass("C6H12O6", which = "exact_mass")

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
