% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_check_functions.R
\name{check_adduct_formula}
\alias{check_adduct_formula}
\title{Validate if a given adduct formula is correct}
\usage{
check_adduct_formula(adduct)
}
\arguments{
\item{adduct}{A character string representing the adduct formula to be validated.}
}
\value{
A logical value: TRUE if the adduct formula is valid, FALSE otherwise.
}
\description{
This function checks whether a given adduct formula is valid by comparing it to known valid adducts.
It returns TRUE if the adduct formula is valid, otherwise returns FALSE.
}
\examples{
check_adduct_formula("(M+H)+") # TRUE
check_adduct_formula("(M+X)+") # FALSE

}
