% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_check_functions.R
\name{check_chemical_formula}
\alias{check_chemical_formula}
\title{Validate if a given chemical formula is a standard valid formula}
\usage{
check_chemical_formula(formula)
}
\arguments{
\item{formula}{A character string representing the chemical formula to be validated.}
}
\value{
A logical value: TRUE if the formula is valid, FALSE otherwise.
}
\description{
This function checks whether a given chemical formula is valid based on standard conventions.
It ensures that element symbols are followed by valid numeric subscripts (if present).
The function uses a list of known chemical elements up to 2022.
It returns TRUE if the formula is valid, otherwise returns FALSE.
}
\examples{
check_chemical_formula("H2O") # TRUE
check_chemical_formula("2H") # FALSE

}
