% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_formula.R
\name{convert_precursor_mz2accurate_mass}
\alias{convert_precursor_mz2accurate_mass}
\title{convert_precursor_mz2accurate_mass}
\usage{
convert_precursor_mz2accurate_mass(precursor_mz = 805.559, adduct = "2M-2H+Na")
}
\arguments{
\item{precursor_mz}{Chemical formula.}

\item{adduct}{Adduct.}
}
\value{
formula
}
\description{
convert_precursor_mz2accurate_mass
}
\examples{
convert_precursor_mz2accurate_mass(
precursor_mz = 805.559,
adduct = "2M-2H+Na")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
