% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classyfire.R
\name{get_compound_class}
\alias{get_compound_class}
\title{get_compound_class}
\usage{
get_compound_class(
  inchikey = "QZDWODWEESGPLC-UHFFFAOYSA-N",
  server = "http://classyfire.wishartlab.com/entities/",
  sleep = 5
)
}
\arguments{
\item{inchikey}{The inchikey ID of a metabolite.}

\item{server}{server.}

\item{sleep}{Sleep time for system. Second.}
}
\value{
A classyfire class object.
}
\description{
Get the class information of a compound using classyfire.
}
\examples{
get_compound_class()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
