% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_functions.R
\name{get_dp}
\alias{get_dp}
\title{get_dp}
\usage{
get_dp(exp.int, lib.int)
}
\arguments{
\item{exp.int}{exp.int}

\item{lib.int}{lib.int}
}
\value{
dot product
}
\description{
get_dp
}
\examples{
get_dp(exp.int = 1:10, lib.int = 1:10)
}
