% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_tools.R
\name{get_os}
\alias{get_os}
\title{Determine the Operating System of the Current R Session}
\usage{
get_os()
}
\value{
A character string indicating the operating system. Possible values are "windows", "osx", "linux", or others based on the system.
}
\description{
This function identifies the operating system on which the R session is running and returns it in a standardized format.
}
\examples{
get_os()
}
