% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_tools.R
\name{get_os}
\alias{get_os}
\title{get_os}
\usage{
get_os()
}
\value{
os name
}
\description{
Convert from wide data to long data.
}
\examples{
get_os()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
