% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_wd.R
\name{get_project_wd}
\alias{get_project_wd}
\title{get_project_wd}
\usage{
get_project_wd()
}
\value{
A working directory.
}
\description{
Get the working directory where Rproj object is in.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
