% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_functions.R
\name{get_spectra_match_score}
\alias{get_spectra_match_score}
\title{get_spectra_match_score}
\usage{
get_spectra_match_score(
  exp.spectrum,
  lib.spectrum,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  fraction.weight = 0.2,
  dp.forward.weight = 0.7,
  dp.reverse.weight = 0.1,
  remove_fragment_intensity_coutoff = 0
)
}
\arguments{
\item{exp.spectrum}{exp.spectrum}

\item{lib.spectrum}{lib.spectrum}

\item{ppm.tol}{ppm.tol}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{fraction.weight}{fraction.weight}

\item{dp.forward.weight}{dp.forward.weight}

\item{dp.reverse.weight}{dp.reverse.weight}

\item{remove_fragment_intensity_coutoff}{default is 0 (100 percentage).}
}
\value{
spectrum match score
}
\description{
get_spectra_match_score is used to get two 
MS2 spectra match score, see MS-DIAL
}
\examples{
exp.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
lib.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
get_spectra_match_score(exp.spectrum, lib.spectrum)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
