% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDP}
\alias{getDP}
\title{getDP}
\usage{
getDP(exp.int, lib.int)
}
\arguments{
\item{exp.int}{exp.int}

\item{lib.int}{lib.int}
}
\value{
dot product
}
\description{
getDP
}
\examples{
getDP(exp.int = 1:10, lib.int = 1:10)
}
