% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpectraMatchScore}
\alias{getSpectraMatchScore}
\title{getSpectraMatchScore}
\usage{
getSpectraMatchScore(
  exp.spectrum,
  lib.spectrum,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  fraction.weight = 0.2,
  dp.forward.weight = 0.7,
  dp.reverse.weight = 0.1
)
}
\arguments{
\item{exp.spectrum}{exp.spectrum}

\item{lib.spectrum}{lib.spectrum}

\item{ppm.tol}{ppm.tol}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{fraction.weight}{fraction.weight}

\item{dp.forward.weight}{dp.forward.weight}

\item{dp.reverse.weight}{dp.reverse.weight}
}
\value{
spectrum match score
}
\description{
getSpectraMatchScore
}
\examples{
exp.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
lib.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
getSpectraMatchScore(exp.spectrum, lib.spectrum)
}
