% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-utils.R
\name{install_fastgit}
\alias{install_fastgit}
\title{install_fastgit}
\usage{
install_fastgit(pkg, from = c("gitee", "gitlab", "github"), ...)
}
\arguments{
\item{pkg}{pkg name from github, gitlab or gitee, "name/repo" format}

\item{from}{gitlab, github or gitee.}

\item{...}{Other parameters for install_git}
}
\description{
install packages from fastgit. Credit to Shixiang Wang
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
