% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_rt_match.R
\name{keep_one}
\alias{keep_one}
\title{keep_one}
\usage{
keep_one(result, according.to = c("mz.error", "rt.error"))
}
\arguments{
\item{result}{result from mz_rt_match function.}

\item{according.to}{According to mz error or rt error?}
}
\value{
Return a result without multiple vs. one.
}
\description{
Remove multiple vs. one in result from mz_rt_match function.
}
\examples{
data1 <- data.frame(mz = 1:10, rt = 1:10)
data2 <- data.frame(mz = 1:10, rt = 1:10)
result <- mz_rt_match(data1, data2, mz.tol = 10)

keep_one(result)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
