% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytools_logo.R
\name{masstools_logo}
\alias{masstools_logo}
\title{masstools_logo}
\usage{
masstools_logo()
}
\value{
masstools logo
}
\description{
Get the detailed of metPath package.
}
\examples{
masstools_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
