% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{masstools_packages}
\alias{masstools_packages}
\title{List all packages in the masstools}
\usage{
masstools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include masstools in the list?}
}
\value{
masstools packages
}
\description{
List all packages in the masstools
}
\examples{
masstools_packages()
}
