% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Spectra.R
\name{mergeMS}
\alias{mergeMS}
\title{mergeMS}
\usage{
mergeMS(speclist, ppm = 5, mzdiff = 5e-04)
}
\arguments{
\item{speclist}{list of spectra (matrix with mz and intensity values)}

\item{ppm}{min difference between peaks in ppm}

\item{mzdiff}{min difference between peaks in m/z}
}
\description{
Merge MS spectra, best used with noise-free spectra. Will merge all peaks that are EITHER within ppm OR within mzdiff range (mzdiff typically used to allow larger ppm difference in low mz range)
}
\details{
Replacement for mergeMS(..., mergeOnly = T)
}
