% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Spectra.R
\name{mergeMS}
\alias{mergeMS}
\title{mergeMS}
\usage{
mergeMS(speclist, ppm = 5, mzdiff = 5e-04, count = FALSE,
  iterative = T, removeZeros = T, noiselevel = 0, maxpeaks = NULL,
  toleranceFactor = 1)
}
\arguments{
\item{speclist}{list of spectra (matrices, all with two columns: \code{mz} and \code{intensity})}

\item{ppm}{min difference between peaks in ppm}

\item{mzdiff}{min difference between peaks in m/z}

\item{count}{if TRUE, will count how many peaks were combined into a peak, and add a \code{count} column in the resulting spectrum}

\item{iterative}{if TRUE, will iteratively merge two adjacent peaks within tolerance, 
and then check if there are peaks within tolerance of merged peak}

\item{removeZeros}{remove all entries with 0 intensity}

\item{noiselevel}{all peaks below this intensity (relative to 
highest peak in merged spectrum at relative intensity 1) will be removed}

\item{maxpeaks}{if not NULL, maximum number of peaks in merged spectrum.
Lowest intensity peaks will be removed from merged spectrum.}

\item{toleranceFactor}{should be a numeric vector of increasing values to recursively
apply as a factor to the ppm and mzdiff tolerances.}
}
\value{
if \code{speclist} is a \code{Spectra} object or
 a \code{list} of \code{Spectrum} objects: returns a \code{Spectrum} object.
if \code{speclist} is a \code{matrix} object or
 a \code{list} of \code{matrix} objects: returns a \code{Spectrum} object.
}
\description{
Merge MS spectra, best used with noise-free spectra. Will merge all peaks that are EITHER within ppm OR within mzdiff range (mzdiff typically used to allow larger ppm difference in low mz range)
}
\details{
Replacement for mergeMS(..., mergeOnly = T)
}
