% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ms2_match}
\alias{ms2_match}
\title{ms2_match}
\usage{
ms2_match(exp.spectrum, lib.spectrum, ppm.tol = 30, mz.ppm.thr = 400)
}
\arguments{
\item{exp.spectrum}{exp.spectrum}

\item{lib.spectrum}{lib.spectrum}

\item{ppm.tol}{ppm.tol}

\item{mz.ppm.thr}{mz.ppm.thr}
}
\value{
ms2 match data frame
}
\description{
ms2_match
}
\examples{
exp.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
lib.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
ms2_match(exp.spectrum, lib.spectrum)
}
