% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_plot.R
\name{ms2_plot}
\alias{ms2_plot}
\title{ms2_plot2}
\usage{
ms2_plot(
  spectrum1,
  spectrum2,
  spectrum1_name = "spectrum1",
  spectrum2_name = "spectrum2",
  range.mz,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity",
  col1 = "red",
  col2 = "black",
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  legend.title.size = 15,
  legend.text.size = 15,
  interactive_plot = FALSE
)
}
\arguments{
\item{spectrum1}{Spectrum 1.}

\item{spectrum2}{Spectrum 2.}

\item{spectrum1_name}{Name of spectrum1}

\item{spectrum2_name}{Name of spectrum2}

\item{range.mz}{range.mz}

\item{ppm.tol}{ppm.tol}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{col1}{Color 1.}

\item{col2}{Color 2.}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size.}

\item{legend.title.size}{legend.title.size}

\item{legend.text.size}{legend.text.size}

\item{interactive_plot}{Interactive plot or not.}
}
\value{
Return a MS2 spectrum.
}
\description{
Plot ms2 spectra or ms2 match plot.
}
\examples{
spectrum1 <- data.frame(
    mz = c(
        87.50874,
        94.85532,
        97.17808,
        97.25629,
        103.36186,
        106.96647,
        107.21461,
        111.00887,
        113.79269,
        118.70564
    ),
    intensity =
        c(
            8356.306,
            7654.128,
            9456.207,
            8837.188,
            8560.228,
            8746.359,
            8379.361,
            169741.797,
            7953.080,
            8378.066
        )
)
spectrum2 <- spectrum1
ms2_plot(spectrum1, spectrum2)
ms2_plot(spectrum1, spectrum2, interactive_plot = TRUE)
ms2_plot(spectrum1)
ms2_plot(spectrum1, interactive_plot = TRUE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
