% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_rt_match.R
\name{mz_rt_match}
\alias{mz_rt_match}
\title{mz_rt_match}
\usage{
mz_rt_match(
  data1,
  data2,
  mz.tol,
  rt.tol = 30,
  rt.error.type = c("relative", "abs")
)
}
\arguments{
\item{data1}{First data for matching, first column must be mz
and seconod column must be rt.}

\item{data2}{Second data for matching, first column must be mz
and seconod column must be rt.}

\item{mz.tol}{mz tol for ms1 and ms2 data matching.}

\item{rt.tol}{RT tol for ms1 and ms2 data matching.}

\item{rt.error.type}{RT error is calculated with relative or absolute.}
}
\value{
Return a result which give the matching
result of data1 and database.
}
\description{
Match peaks according to m/z and RT.
}
\examples{
data1 <- data.frame(mz = 1:10, rt = 1:10)
data2 <- data.frame(mz = 1:10, rt = 1:10)
mz_rt_match(data1, data2, mz.tol = 10)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
