% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_duplicated.R
\name{name_duplicated}
\alias{name_duplicated}
\title{Handle Duplicated Names by Appending Sequence Number}
\usage{
name_duplicated(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector where duplicated names are made unique by appending a sequence number.
}
\description{
This function checks for duplicated elements in a vector. If duplicated elements are found,
it appends an underscore and a sequence number to each occurrence of the duplicated element.
}
\examples{
vec <- c("apple", "orange", "apple", "banana", "orange")
name_duplicated(vec)

}
