% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_duplicated.R
\name{name_duplicated}
\alias{name_duplicated}
\title{name_duplicated}
\usage{
name_duplicated(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector without duplicated.
}
\description{
name_duplicated
}
\examples{
name_duplicated(c("a", "a", "b", "c", "a", "b", "c", "a"))
name_duplicated(c(rep(1, 5), 2))
name_duplicated(1:5)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
