% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_MFobject.R
\name{Ops.MFobject}
\alias{Ops.MFobject}
\title{Ops.MFobject}
\usage{
\method{Ops}{MFobject}(e1, e2)
}
\arguments{
\item{e1}{numeric or MFobject}

\item{e2}{numeric or MFobject}
}
\description{
Operator behavior for MFobject S3 class
}
