% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_chemical_formula_funcitons.R
\name{parse_chemical_formula}
\alias{parse_chemical_formula}
\title{Parse a Chemical Formula into its Constituent Elements and Counts}
\usage{
parse_chemical_formula(formula)
}
\arguments{
\item{formula}{Character string representing the chemical formula to be parsed.}
}
\value{
A data frame with two columns: 'Element' and 'Count', representing each element and its count in the given formula.
}
\description{
Breaks down a chemical formula into its individual elements and counts,
returning the results as a data frame.
}
\details{
Parse a Chemical Formula into its Constituent Elements and Counts

This function breaks down a chemical formula into its individual elements and their respective counts.
If no count is specified for an element, it defaults to 1.
The result is presented as a data frame with columns for elements and counts.
}
\examples{
parse_chemical_formula("H2O")       # Returns a data frame with H:2 and O:1
parse_chemical_formula("C6H12O6")   # Returns a data frame with C:6, H:12, and O:6
parse_chemical_formula("Fe2O3")     # Returns a data frame with Fe:2 and O:3

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
