% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgf.R
\name{read_mgf}
\alias{read_mgf}
\title{Read and Process MGF Files}
\usage{
read_mgf(file)
}
\arguments{
\item{file}{A character vector specifying the path(s) to the MGF file(s).}
}
\value{
A list containing processed MS2 spectra information for each provided MGF file. 
Each element of the list contains two components: 
\itemize{
  \item \code{info}: A named numeric vector containing the m/z and retention time of the precursor ion.
  \item \code{spec}: A matrix where each row represents a fragment ion peak, with columns for m/z and intensity values.
}
Empty spectra are removed from the output.
}
\description{
This function reads MGF files and extracts the relevant MS2 spectra information. 
Duplicated peaks and noise can be optionally removed.
}
\examples{
# Locate the example mgf file in the installed package
file_path <- system.file("extdata", "example.mgf", package = "masstools")
# Then use it in your function
result <- read_mgf(file_path)
}
