% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgf4database.R
\name{read_mgf4database}
\alias{read_mgf4database}
\title{read_mgf4database}
\usage{
read_mgf4database(file)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mgf.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MGF data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
