% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzxml.R
\name{read_mzxml}
\alias{read_mzxml}
\title{Read and Process mzXML Files}
\usage{
read_mzxml(file, threads = 3, mode = c("inMemory", "onDisk"))
}
\arguments{
\item{file}{A character vector specifying the path(s) to the mzXML file(s).}

\item{threads}{An integer, indicating the number of cores to use (default: 3).}

\item{mode}{A character string, either \code{"inMemory"} or \code{"onDisk"} 
specifying the mode of reading data. Default is \code{"inMemory"}.}
}
\value{
A list containing processed MS2 spectra information for each provided mzXML file. 
Each element of the list contains two components: 
\itemize{
  \item \code{info}: A data frame with columns for the name (composed of m/z and retention time), 
  m/z, retention time, and file name.
  \item \code{spec}: A data frame where each row represents a fragment ion peak, 
  with columns for m/z and intensity values.
}
}
\description{
Reads mzXML files and extracts relevant MS2 spectra information.
}
\examples{
# Assuming 'sample.mzXML' is the path to your mzXML file.
# mzxml_data <- read_mzxml('sample.mzXML')

}
