% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_noise}
\alias{remove_noise}
\title{remove_noise}
\usage{
remove_noise(spec, ppm.ms2match = 30, mz.ppm.thr = 400)
}
\arguments{
\item{spec}{spec}

\item{ppm.ms2match}{ppm.ms2match}

\item{mz.ppm.thr}{mz.ppm.thr}
}
\value{
clear spec
}
\description{
remove_noise
}
\examples{
exp.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
remove_noise(exp.spectrum)
}
