% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeNoise}
\alias{removeNoise}
\title{removeNoise}
\usage{
removeNoise(spec, ppm.ms2match = 30, mz.ppm.thr = 400)
}
\arguments{
\item{spec}{spec}

\item{ppm.ms2match}{ppm.ms2match}

\item{mz.ppm.thr}{mz.ppm.thr}
}
\value{
clear spec
}
\description{
removeNoise
}
\examples{
exp.spectrum <- data.frame(mz = 1:10, intensity = 1:10)
removeNoise(exp.spectrum)
}
