% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classyfire.R
\name{show,classyfire-method}
\alias{show,classyfire-method}
\title{Display Method for 'classyfire' Objects}
\usage{
\S4method{show}{classyfire}(object)
}
\arguments{
\item{object}{An object of class 'classyfire'.}
}
\description{
This function provides a custom display for 'classyfire' S4 objects.
It prints out the main details of the object including compound 
information and its taxonomic classification in a structured format.
}
\examples{
\dontrun{
cf <- new("classyfire", ...)
show(cf)
}

}
\author{
Xiaotao Shen <shenxt1990@outlook.com>
}
