% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-chemical_formula_funcitons.R
\name{split_formula}
\alias{split_formula}
\title{split_formula}
\usage{
split_formula(formula = "C9H11NO2")
}
\arguments{
\item{formula}{Chemical formula.}
}
\value{
A spited formula.
}
\description{
Split a formula into element and number.
}
\examples{
split_formula(formula = "C9H11NO2")
split_formula(formula = "CH3")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
