% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_deprecated_functions.R
\name{split_formula}
\alias{split_formula}
\title{Split Chemical Formula}
\usage{
split_formula(formula = "C9H11NO2")
}
\arguments{
\item{formula}{A character string representing a chemical formula (e.g., "C9H11NO2").
Default is "C9H11NO2".}
}
\value{
A data frame with two columns: `element` contains the elemental symbols
and `number` contains the respective counts for each element in the formula.
}
\description{
Takes a chemical formula string and breaks it down into its
individual elemental components and their respective counts.
}
\examples{
split_formula(formula = "C9H11NO2")
split_formula(formula = "CH3")
split_formula(formula = "CH3C")
}
