% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_chemical_formula_funcitons.R
\name{sum_formula}
\alias{sum_formula}
\title{Summation of Chemical Formulas with Adducts}
\usage{
sum_formula(formula = "C9H11NO2", adduct = "M-H2O+H")
}
\arguments{
\item{formula}{A character string representing the base chemical formula.
Default is "C9H11NO2".}

\item{adduct}{A character string representing the adduct to be added or
subtracted from the base formula. Examples include "M-H2O+H", "M+", and "M-".
Default is "M-H2O+H".}
}
\value{
A character string of the resulting chemical formula after combining
the base formula with the adduct. If an error occurs or the summation results
in negative counts for any element, it returns `NA`.
}
\description{
Combines a chemical formula with a specified adduct, and returns the resultant
summed formula. This function can handle addition or subtraction of elements
from the main formula based on the adduct.
}
\examples{
sum_formula(formula = "C9H11NO2", adduct = "M+H")
sum_formula(formula = "C9H11NO2", adduct = "M+")
sum_formula(formula = "C9H11NO2", adduct = "M+CH3COOH")
sum_formula(formula = "C9H11", adduct = "M-H20")
}
\author{
Xiaotao Shen <xiaotao.shen@outlook.com>
}
