% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_formula.R
\name{sum_formula}
\alias{sum_formula}
\title{sum_formula}
\usage{
sum_formula(formula = "C9H11NO2", adduct = "M-H2O+H")
}
\arguments{
\item{formula}{Chemical formula.}

\item{adduct}{Adduct.}
}
\value{
formula
}
\description{
Get the total formula after add adduct.
}
\examples{
sum_formula(formula = "C9H11NO2", adduct = "M+H")
sum_formula(formula = "C9H11NO2", adduct = "M+")
sum_formula(formula = "C9H11NO2", adduct = "M+CH3COOH")
sum_formula(formula = "C9H11", adduct = "M-H20")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
