% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_ID.R
\name{trans_ID}
\alias{trans_ID}
\title{Convert Chemical Identifier Between Different Databases}
\usage{
trans_ID(
  query = "C00001",
  from = "KEGG",
  to = "PubChem SID",
  top = 1,
  server = c("cts.fiehnlab", "chemspider")
)
}
\arguments{
\item{query}{A character string of the chemical identifier to be converted.}

\item{from}{A character string of the database from which the query is.}

\item{to}{A character string of the desired database to which the identifier should be converted.}

\item{top}{Numeric value indicating the top results to be returned.}

\item{server}{A character string indicating which server to use. Possible values are "cts.fiehnlab" or "chemspider".}
}
\value{
A character string of the converted identifier, or `NA` if the conversion was unsuccessful.
}
\description{
This function takes an identifier from a given database (`from`) and converts it to the desired database (`to`).
It uses either the `cts.fiehnlab` or `chemspider` services for conversion.
}
\examples{
\dontrun{
trans_ID(
    query = "C00001",
    from = "KEGG",
    to = "PubChem SID",
    top = 1,
    server = "cts.fiehnlab"
)

trans_ID(
    query = "C00001",
    to = "Human Metabolome Database",
    server = "cts.fiehnlab"
)
}
}
