% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_ID.R
\name{trans_id_database}
\alias{trans_id_database}
\title{trans_id_database}
\usage{
trans_id_database(server = c("cts.fiehnlab", "chemspider"))
}
\arguments{
\item{server}{server. cts.fiehnlab
(http://cts.fiehnlab.ucdavis.edu/service/convert)
or chemspider (https://www.chemspider.com/InChI.asmx)}
}
\value{
A vector..
}
\description{
Whate databases are supported.
}
\examples{
trans_id_database(server = "cts.fiehnlab")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
